requirements.txt
asgiref
aiohttp
bidict==0.22.1
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
dnspython==2.4.2
eventlet==0.33.3
fire==0.5.0
Flask==3.0.0
Flask-SocketIO==5.3.6
greenlet==3.0.0
h11==0.14.0
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
loguru==0.7.2
MarkupSafe==2.1.3
python-engineio==4.7.1
PyYAML==6.0.1
requests==2.31.0
simple-websocket==0.10.1
six==1.16.0
termcolor==2.3.0
urllib3==2.0.6
websocket-client==1.6.3
Werkzeug==3.0.0
wsproto==1.2.0
zipp==3.17.0
requests
docker